config: move `DyanmicConfig` to `sampler`
